+++ R/have-motif.R
@@ -466,13 +466,12 @@ have_motif_ <- function(
   }
 
   if (
-    mode == "strict" &&
-      !composition_can_match(
-        glycan_graph,
-        motif_composition_profile,
-        glycan_batch_profile = glycan_batch_profile,
-        motif_batch_profile = motif_batch_profile
-      )
+    !composition_can_match(
+      glycan_graph,
+      motif_composition_profile,
+      glycan_batch_profile = glycan_batch_profile,
+      motif_batch_profile = motif_batch_profile
+    )
   ) {
     return(FALSE)
   }